LPS-65385 Move DestinationNames to com.liferay.exportimport.resources.importer.constants
